Disable `minimumReleaseAge` for Biome
`customManagers:biomeVersions` does not respect the minimum release age
for the corresponding npm package of Biome, causing Renovate to open a
pull request that upgrades the `$schema` property in `biome.json`
without upgrading Biome itself.
This commit disables `minimumReleaseAge` for the Biome npm package so
that it is upgraded along with the `$schema` property in `biome.json`.