Revise the GitHub Actions workflows
Specifying the `name` and `version` properties in `package.json` is
necessary to avoid triggering the `publish-tags.yml` workflow
unnecessarily. Otherwise, the workflow would be triggered when any
branch is merged to main. While it skips the 'tags' job for non-release
branches, it still clutters the workflow run overview in GitHub Actions.