Omit the version of type definition upgrades
The `@types/**` package group is an umbrella group that covers all type
definitions. Renovate automatically omits the version number from the
commit message when it has to upgrade multiple type definition packages.
This commit makes it omit the version number when it has to upgrade only
a single type definition package.